+++ publish_skate.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+./gradlew :skate-plugin:uploadPluginToArtifactory --no-configuration-cache --stacktrace