+++ skate-plugin/src/main/kotlin/com/slack/sgp/intellij/ui/WhatsNewPanelFactory.kt
@@ -18,6 +18,7 @@ package com.slack.sgp.intellij.ui
 import com.intellij.openapi.Disposable
 import com.intellij.openapi.application.runReadAction
 import com.intellij.openapi.components.service
+import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.DumbAware
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.Disposer
@@ -26,12 +27,17 @@ import com.intellij.testFramework.LightVirtualFile
 import com.intellij.ui.content.ContentFactory
 import com.intellij.ui.content.ContentManagerEvent
 import com.intellij.ui.content.ContentManagerListener
+import com.intellij.ui.dsl.builder.panel
+import com.intellij.ui.dsl.gridLayout.HorizontalAlign
+import com.intellij.ui.dsl.gridLayout.VerticalAlign
+import com.intellij.ui.jcef.JBCefApp
+import com.intellij.util.ui.HtmlPanel
+import com.intellij.util.ui.JBUI
 import com.slack.sgp.intellij.ChangelogJournal
 import com.slack.sgp.intellij.ChangelogParser
 import java.awt.BorderLayout
 import javax.swing.JComponent
 import javax.swing.JPanel
-import org.intellij.lang.annotations.Language
 import org.intellij.plugins.markdown.ui.preview.html.MarkdownUtil
 import org.intellij.plugins.markdown.ui.preview.jcef.MarkdownJCEFHtmlPanel
 
@@ -46,11 +52,11 @@ class WhatsNewPanelFactory : DumbAware {
   fun createToolWindowContent(
     toolWindow: ToolWindow,
     project: Project,
-    markdownFileString: String,
+    changeLogContent: ChangelogParser.PresentedChangelog,
     parentDisposable: Disposable
   ) {
 
-    val toolWindowContent = WhatsNewPanelContent(project, markdownFileString, parentDisposable)
+    val toolWindowContent = WhatsNewPanelContent(project, changeLogContent, parentDisposable)
     val content =
       ContentFactory.getInstance().createContent(toolWindowContent.contentPanel, "", false)
     toolWindow.contentManager.addContent(content)
@@ -68,44 +74,75 @@ class WhatsNewPanelFactory : DumbAware {
 
   private class WhatsNewPanelContent(
     project: Project,
-    @Language("Markdown") markdownFileString: String,
+    changeLogContent: ChangelogParser.PresentedChangelog,
     parentDisposable: Disposable
   ) {
 
-    // Actual panel box for What's New at Slack
+    private val logger = logger<WhatsNewPanelContent>()
+
+    // Actual panel box for "What's New in Slack!"
     val contentPanel: JPanel =
       JPanel().apply {
         layout = BorderLayout(0, 20)
-        add(createWhatsNewPanel(project, markdownFileString, parentDisposable), BorderLayout.CENTER)
+        add(createWhatsNewPanel(project, changeLogContent, parentDisposable), BorderLayout.CENTER)
       }
 
-    // Control Panel that takes in the current project, markdown string, and a Disposable.
+    // Control Panel that takes in the current project, parsed string, and a Disposable.
     private fun createWhatsNewPanel(
       project: Project,
-      @Language("Markdown") markdownFileString: String,
+      changeLogContent: ChangelogParser.PresentedChangelog,
       parentDisposable: Disposable
     ): JComponent {
       // to take in the parsed Changelog:
       val changelogJournal = project.service<ChangelogJournal>()
 
-      val parsedChangelog =
-        ChangelogParser.readFile(markdownFileString, changelogJournal.lastReadDate)
+      changelogJournal.lastReadDate = changeLogContent.lastReadDate
 
-      changelogJournal.lastReadDate = parsedChangelog.lastReadDate
+      val file = LightVirtualFile("changelog.md", changeLogContent.changeLogString ?: "")
 
-      // then, pass in parsedChangelog instead of markdownFileString
-      val file = LightVirtualFile("changelog.md", parsedChangelog.changeLogString ?: "")
-
-      val panel = MarkdownJCEFHtmlPanel(project, file)
-      Disposer.register(parentDisposable, panel)
-
-      // TODO: if it's empty, don't show the panel
       val html = runReadAction {
-        MarkdownUtil.generateMarkdownHtml(file, parsedChangelog.changeLogString ?: "", project)
+        MarkdownUtil.generateMarkdownHtml(file, changeLogContent.changeLogString ?: "", project)
       }
 
-      panel.setHtml(html, 0)
-      return panel.component
+      // We have to support two different modes here: the modern JCEF-based one and the legacy
+      // Legacy is because Android Studio ships with a broken markdown plugin that doesn't work with
+      // JCEF
+      // https://issuetracker.google.com/issues/159933628#comment19
+      val panel =
+        if (JBCefApp.isSupported()) {
+          logger.debug("Using JCEFHtmlPanelProvider")
+          MarkdownJCEFHtmlPanel(project, file)
+            .apply {
+              Disposer.register(parentDisposable, this)
+              setHtml(html, 0)
+            }
+            .component
+        } else {
+          logger.debug("Using HtmlPanel")
+          val htmlPanel =
+            object : HtmlPanel() {
+              init {
+                isVisible = true
+                // Padding
+                border = JBUI.Borders.empty(16)
+                update()
+              }
+
+              override fun getBody(): String {
+                return html
+              }
+            }
+          panel {
+            row {
+                scrollCell(htmlPanel)
+                  .horizontalAlign(HorizontalAlign.FILL)
+                  .verticalAlign(VerticalAlign.FILL)
+              }
+              .resizableRow()
+          }
+        }
+
+      return panel
     }
   }
 }